refactor: split pack schema facade
