+++ crates/assay-evidence/src/lint/packs/schema_next/mod.rs
@@ -1,0 +1,12 @@
+pub(crate) mod conditional;
+pub(crate) mod errors;
+pub(crate) mod serde;
+pub(crate) mod types;
+pub(crate) mod validation;
+
+pub use errors::PackValidationError;
+pub use types::{
+    CheckDefinition, PackDefinition, PackKind, PackRequirements, PackRule,
+    SupportedConditionalCheck, SupportedConditionalClause,
+};
+pub use validation::is_valid_pack_name;
